feat(header): move header.js component to new folder header
