+++ src/components/layout/layout.js
@@ -1,25 +1,26 @@
 import PropTypes from "prop-types";
 import React from "react";
 import Footer from "../footer/footer";
-import Header from "../header";
+import Header from "../header/header";
 import styles from "./layout.module.css";
-import pathImg from '../../images/background/path.svg';
-import BlogSection from '../blog-section/blog-section';
+import pathImg from "../../images/background/path.svg";
+import BlogSection from "../blog-section/blog-section";
 import Subscribe from "../subscribe/subscribe";
 
 const Layout = ({ isHome, children }) => {
-  let background = ''
-  let blogSection = null
+  let background = "";
+  let blogSection = null;
 
   if (isHome) {
-    background = pathImg
-    blogSection = <BlogSection />
+    background = pathImg;
+    blogSection = <BlogSection />;
   }
 
   return (
     <div
       className={styles.layout}
-      style={{ backgroundImage: `url(${background})` }}>
+      style={{ backgroundImage: `url(${background})` }}
+    >
       <Header />
       <main className={styles.mainContent}>{children}</main>
       <Subscribe />
@@ -27,7 +28,7 @@ const Layout = ({ isHome, children }) => {
       <Footer />
     </div>
   );
-}
+};
 
 Layout.propTypes = {
   isHome: PropTypes.bool, // Indicates whether it is the Homepage or not
@@ -35,7 +36,7 @@ Layout.propTypes = {
 };
 
 Layout.defaultProps = {
-  isHome: false
-}
+  isHome: false,
+};
 
 export default Layout;
